Update first.js
calculatSQRT is not defined error was found because I didn't add this function in the DOM. then a few things are added for validation purpose. a new variable errroText is made to show an error message in every validation if else condition.  in makeExp function, an if-else condition is used to check if the user enters any of the operators twice.then it shows an error message.